# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agro API",
        "description": "Documentación de la API de autenticación con Swagger en Laravel 11",
        "contact": {
            "email": "soporte@tuempresa.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/info-contact": {
            "get": {
                "tags": [
                    "InfoContact"
                ],
                "summary": "Obtener la información de contacto",
                "operationId": "b77438071e605eba8ebe76a637b8bd99",
                "responses": {
                    "200": {
                        "description": "Información de contacto obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Información no encontrada"
                    }
                }
            }
        },
        "/api/info-contact/{id}": {
            "put": {
                "tags": [
                    "InfoContact"
                ],
                "summary": "Actualizar la información de contacto",
                "operationId": "4550cbd3863ce668d6147963f829e7fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la información de contacto a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "location": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Lima, Perú"
                                    },
                                    "cellphone": {
                                        "type": "string",
                                        "maxLength": 9,
                                        "minLength": 9,
                                        "example": "987654321"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 320,
                                        "example": "contacto@empresa.com"
                                    },
                                    "attention_hours": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Lunes a Viernes de 9am a 6pm"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Información de contacto actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Información de contacto actualizada correctamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "location": {
                                                    "type": "string",
                                                    "example": "Lima, Perú"
                                                },
                                                "cellphone": {
                                                    "type": "string",
                                                    "example": "987654321"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "contacto@empresa.com"
                                                },
                                                "attention_hours": {
                                                    "type": "string",
                                                    "example": "Lunes a Viernes de 9am a 6pm"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "No se enviaron datos válidos para actualizar",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No se enviaron datos válidos para actualizar."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Información de contacto no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Información de contacto no encontrada."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string",
                                                    "example": "El campo location es obligatorio."
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{nameProduct}/pdf": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Actualizar el PDF de un producto",
                "description": "Elimina el PDF anterior de un producto y guarda uno nuevo.",
                "operationId": "d1f2004e41118d8161b27e11721abe4f",
                "parameters": [
                    {
                        "name": "nameProduct",
                        "in": "path",
                        "description": "Nombre del producto cuyo PDF se actualizará.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pdf"
                                ],
                                "properties": {
                                    "pdf": {
                                        "description": "Archivo PDF en formato Base64.",
                                        "type": "string",
                                        "format": "base64"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "PDF actualizado correctamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "Pdf actualizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Producto no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la solicitud.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El archivo PDF es inválido"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}